+++ packages/perseus-core/src/parse-perseus-json/regression-tests/item-data/radio-choice-v2-with-rationale.ts
@@ -1,0 +1,250 @@
+// WARNING: Do not change or delete this file! If you do, Perseus might become
+// unable to parse the current data format, which will break clients.
+// If you need to add more regression tests, add a new file to this directory.
+export default {
+    answerArea: {
+        calculator: false,
+        options: {
+            content: "",
+            images: {},
+            widgets: {},
+        },
+        type: "multiple",
+    },
+    hints: [],
+    itemDataVersion: {
+        major: 0,
+        minor: 1,
+    },
+    question: {
+        content:
+            "\n\n[[☃ group 1]]\n\n[[☃ group 2]]\n\n[[☃ group 3]]\n\n[[☃ group 4]]",
+        images: {},
+        widgets: {
+            "group 1": {
+                graded: true,
+                options: {
+                    content: "\n\n[[☃ radio 1]]",
+                    images: {},
+                    widgets: {
+                        "radio 1": {
+                            graded: true,
+                            options: {
+                                choices: [
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                ],
+                                deselectEnabled: false,
+                                displayCount: null,
+                                multipleSelect: false,
+                                noneOfTheAbove: false,
+                                onePerLine: true,
+                                randomize: false,
+                            },
+                            type: "radio",
+                            version: {
+                                major: 2,
+                                minor: 0,
+                            },
+                        },
+                    },
+                },
+                type: "group",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+            "group 2": {
+                graded: true,
+                options: {
+                    content: "\n\n[[☃ radio 1]]",
+                    images: {},
+                    widgets: {
+                        "radio 1": {
+                            graded: true,
+                            options: {
+                                choices: [
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                ],
+                                deselectEnabled: false,
+                                displayCount: null,
+                                multipleSelect: false,
+                                noneOfTheAbove: false,
+                                onePerLine: true,
+                                randomize: false,
+                            },
+                            type: "radio",
+                            version: {
+                                major: 2,
+                                minor: 0,
+                            },
+                        },
+                    },
+                },
+                type: "group",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+            "group 3": {
+                graded: true,
+                options: {
+                    content: "\n\n[[☃ radio 1]]",
+                    images: {},
+                    widgets: {
+                        "radio 1": {
+                            graded: true,
+                            options: {
+                                choices: [
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                ],
+                                deselectEnabled: false,
+                                displayCount: null,
+                                multipleSelect: false,
+                                noneOfTheAbove: false,
+                                onePerLine: true,
+                                randomize: false,
+                            },
+                            type: "radio",
+                            version: {
+                                major: 2,
+                                minor: 0,
+                            },
+                        },
+                    },
+                },
+                type: "group",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+            "group 4": {
+                graded: true,
+                options: {
+                    content: "\n\n[[☃ radio 1]]",
+                    images: {},
+                    widgets: {
+                        "radio 1": {
+                            graded: true,
+                            options: {
+                                choices: [
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                    {
+                                        content: "This is some content",
+                                        rationale: "This is some rationale",
+                                        correct: false,
+                                        isNoneOfTheAbove: false,
+                                    },
+                                ],
+                                deselectEnabled: false,
+                                displayCount: null,
+                                multipleSelect: false,
+                                noneOfTheAbove: false,
+                                onePerLine: true,
+                                randomize: false,
+                            },
+                            type: "radio",
+                            version: {
+                                major: 2,
+                                minor: 0,
+                            },
+                        },
+                    },
+                },
+                type: "group",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+        },
+    },
+};