+++ packages/perseus/src/mixins/widget-prop-denylist.ts
@@ -38,12 +38,10 @@ const denylist = [
     "trackInteraction",
     "keypadElement",
     "linterContext",
-    "isLastUsedWidget",
     "handleUserInput",
     "analytics",
     "showSolutions",
     "reviewMode",
-    "reviewModeRubric",
     "widgetIndex",
 ];