+++ packages/perseus/src/widgets/passage/__tests__/passage.test.tsx
@@ -41,7 +41,6 @@ function renderPassage(
         },
         containerSizeClass: "small",
         findWidgets: (callback) => [],
-        isLastUsedWidget: false,
         onBlur: () => {},
         onChange: () => {},
         handleUserInput: () => {},