+++ packages/perseus/src/widgets/radio/__docs__/multiple-choice-interactions-regression.stories.tsx
@@ -27,7 +27,7 @@ type StoryArgs = {
 >;
 
 export default {
-    title: "Widgets/RadioNew/Visual Regression Tests/Interactive",
+    title: "Widgets/RadioNew/Visual Regression Tests/Interactions",
     component: ServerItemRendererWithDebugUI,
     tags: ["!autodocs"],
     parameters: {